+++ src/signed-data-store/signed-data-store.ts
@@ -8,8 +8,8 @@ import { logger } from '../logger';
 let signedApiStore: Record<AirnodeAddress, Record<TemplateId, LocalSignedData>> = {};
 let pruner: NodeJS.Timeout | undefined;
 
-export const checkMessage = ({ airnode, templateId, timestamp, signature, encodedValue }: SignedData) => {
-  // 'data' is: ethers.utils.defaultAbiCoder.encode(['int256'], [beaconValue]);
+export const verifySignedData = ({ airnode, templateId, timestamp, signature, encodedValue }: SignedData) => {
+  // 'encodedValue' is: ethers.utils.defaultAbiCoder.encode(['int256'], [beaconValue]);
 
   const message = ethers.utils.arrayify(
     ethers.utils.solidityKeccak256(['bytes32', 'uint256', 'bytes'], [templateId, timestamp, encodedValue])
@@ -19,12 +19,10 @@ export const checkMessage = ({ airnode, templateId, timestamp, signature, encode
   return signerAddr !== airnode;
 };
 
-export const checkSignedDataIntegrity = (signedData: SignedData) => {
-  const { airnode, templateId, timestamp, encodedValue } = signedData;
-
-  if (parseInt(timestamp) * 1_000 > Date.now()) {
+const checkTimestamp = ({ timestamp, airnode, encodedValue, templateId }: SignedData) => {
+  if (parseInt(timestamp) * 1_000 > Date.now() + 60 * 60 * 1_000) {
     logger.warn(
-      `Refusing to store sample as timestamp is in the future: (Airnode ${airnode}) (Template ID ${templateId}) (Received timestamp ${new Date(
+      `Refusing to store sample as timestamp is more than one hour in the future: (Airnode ${airnode}) (Template ID ${templateId}) (Received timestamp ${new Date(
         parseInt(timestamp) * 1_000
       ).toLocaleDateString()} vs now ${new Date().toLocaleDateString()}), ${
         BigNumber.from(encodedValue).div(10e10).toNumber() / 10e8
@@ -33,7 +31,27 @@ export const checkSignedDataIntegrity = (signedData: SignedData) => {
     return false;
   }
 
-  if (checkMessage(signedData)) {
+  if (parseInt(timestamp) * 1_000 > Date.now()) {
+    logger.warn(
+      `Sample is in the future, but by less than an hour, therefore storing anyway: (Airnode ${airnode}) (Template ID ${templateId}) (Received timestamp ${new Date(
+        parseInt(timestamp) * 1_000
+      ).toLocaleDateString()} vs now ${new Date().toLocaleDateString()}), ${
+        BigNumber.from(encodedValue).div(10e10).toNumber() / 10e8
+      }`
+    );
+  }
+
+  return true;
+};
+
+export const checkSignedDataIntegrity = (signedData: SignedData) => {
+  const { airnode, templateId, timestamp, encodedValue } = signedData;
+
+  if (!checkTimestamp(signedData)) {
+    return;
+  }
+
+  if (verifySignedData(signedData)) {
     logger.warn(
       `Refusing to store sample as signature does not match: (Airnode ${airnode}) (Template ID ${templateId}) (Received timestamp ${new Date(
         parseInt(timestamp) * 1_000
